+++ src/app/services/types/master-scan.types.ts
