+++ src/app/services/types/scan.types.ts
