+++ src/app/views/scans-landing/scans-landing.component.ts
@@ -1,4 +1,4 @@
-import { AWSListing, AWSSite } from '../../services/sites.types';
+import { AWSListing, AWSSite } from '../../services/types/sites.types';
 import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Observable } from 'rxjs';
 import { SiteListingFacadeService } from '../../services/site-listing-facade.service';
